feat: 카테고리별 게시물 조회
- 카테고리 id를 전달받아 해당 카테고리 게시물을 조회할 수 있는 api
- 카테고리 id가 long으로 잘못 설정되어있어 Integer 타입으로 변경
- null 값을 받을 수 있도록 int 타입이 아닌 Integer 타입을 사용
- null 값을 전달받는 경우에는 전체 게시물 조회

+++ src/main/java/com/example/myreform/domain/board/Board.java
@@ -26,7 +26,7 @@ public class Board extends BaseEntity {
     private User user;
 
     @Column(name = "category_id")
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(length = 45, nullable = false, name = "title") // 제목 길이 체크
     private String title;
@@ -64,7 +64,7 @@ public class Board extends BaseEntity {
     }
 
     @Builder
-    public Board(User user, Long category_id, String title, String contents){
+    public Board(User user, Integer category_id, String title, String contents){
         this.user = user;
         this.categoryId = category_id;
         this.title = title;
